requirements-dev.txt
-r requirements-base.txt

# add packages needed only for development

behave-django==1.4.0
black==19.10b0
django-debug-toolbar==2.2
factory-boy==3.0.1
faker==4.1.2
flake8==3.8.3
ipdb==0.13.3
mypy==0.782
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-flake8==1.0.6
pytest-mock==3.3.1
pytest-xdist==1.32.0
PyYAML==5.3.1
selenium==3.141.0
